server/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
src_paths = ["api"]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 160
multi_line_output = 3
use_parentheses = true

[tool.black]
line-length = 160

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 160
disable = [
    "bad-continuation",
    "bad-whitespace",
    "fixme",
    "invalid-name",
    "missing-docstring",
    "missing-module-docstring",
    "no-member",  # disabling until this issue is resolved: https://github.com/PyCQA/pylint/issues/533
    "too-few-public-methods",
    "too-many-ancestors",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-nested-blocks",
    "too-many-public-methods",
    "too-many-return-statements",
    "too-many-statements",
    "wrong-import-order"
]

[tool.pylint.'FORMAT']
max-branches = 15
max-line-length = 160
max-statements = 60

[tool.pytest.ini_options]
addopts = "--cov api --cov-report term-missing --cov-fail-under 80 --log-level DEBUG"

[tool.poetry]
name = "api"
version = "0.1.0"
authors = ["Alan Quillin"]
description = "Keg Volume Monitor"
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
Flask = "^3.1.0"
Flask-RESTx = "^1.3"
SQLAlchemy-Utils = "^0.41.2"
flask-cors = "^5.0.1"
python = "^3.11"
sqlalchemy = "^2.0.40"
python-dateutil = "^2.9.0"
gevent = "^24.11.1"
alembic = "^1.15.2"
#psycopg2-binary = "^2.9.9"
psycopg2-binary = "2.9.9" #pinned cause there was build errors for 2.9.10
simplejson = "^3.20.1"

[tool.poetry.dev-dependencies]
bandit = "^1.6.2"
black = "^20.8b1"
freezegun = "^0.3.15"
isort = "^5.4.2"
pdbpp = "^0.10.2"
py = "^1.10.0"
pylint = "^2.6.0"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
pytest-docker-tools = "^1.0.0"
requests_mock = "^1.8.0"
responses = "<0.12.1" # remove this one responses fixes their bug
wrapt = "^1.12.1" # required by pylint (badly)